add level

+++ courses/prog_base_3/project/subjectClass.cpp
@@ -6,7 +6,7 @@
 using namespace std;
 using namespace sf;
 
-Subject::Subject(Image &image, float X, float Y,int Width,int Height, string Name){
+Subject::Subject(Image &image, float X, float Y, float Width, float Height, string Name){
 		x = X; y = Y; width = Width; height = Height; name = Name; moveTime = 0;
 		speed = 0; health = 100; speed_x = 0; speed_y = 0;
 		life = true; gravity = false; movement = false;
